Creat basic user serializer
