+++ api/src/common/generics/generic_post_serializers.py
@@ -1,0 +1,33 @@
+from users.serializers import BasicUserSerializer
+from rest_framework import serializers
+from .generic_serializers import GenericSerializer
+
+
+# Serializes a generic resource model
+class GenericPostSerializer(GenericSerializer):
+    creator = BasicUserSerializer(read_only=True)
+    is_liked = serializers.SerializerMethodField()
+    likes = serializers.SerializerMethodField()
+    comments = serializers.SerializerMethodField()
+
+    def get_is_liked(self, instance):
+        return instance.liked_users.filter(id=self.context['request'].user.id).exists()
+
+    def get_likes(self, instance):
+        return instance.liked_users.count()
+
+    def get_comments(self, instance):
+        return instance.comment_set.count()
+
+
+# Serializes a generic comment model
+class GenericCommentSerializer(GenericSerializer):
+    creator = BasicUserSerializer(read_only=True)
+    is_liked = serializers.SerializerMethodField()
+    likes = serializers.SerializerMethodField()
+
+    def get_is_liked(self, instance):
+        return instance.liked_users.filter(id=self.context['request'].user.id).exists()
+
+    def get_likes(self, instance):
+        return instance.liked_users.count()
